Adds `fail_on_nonconvergence` option to `pagerank` to provide pagerank results even on non-convergence (#3639)
closes #3613 

Prior to this PR, `pagerank` will raise a `RuntimeError` if it fails to converge, often because the `max_iter` param is set too small (intentionally or otherwise).  This PR adds the optional paramter `fail_on_nonconvergence` which defaults to `True` (ie. the current behavior to ensure backwards-compatibility) that allows a caller to run `pagerank` and get results even if it did not converge.  When `fail_on_nonconvergence` is `False`, `pagerank` will return a tuple containing the pagerank results and a bool indicating if the results converged or not).

Authors:
  - Rick Ratzel (https://github.com/rlratzel)
  - Chuck Hastings (https://github.com/ChuckHastings)
  - Brad Rees (https://github.com/BradReesWork)

Approvers:
  - Brad Rees (https://github.com/BradReesWork)
  - Vibhu Jawa (https://github.com/VibhuJawa)
  - Naim (https://github.com/naimnv)
  - Seunghwa Kang (https://github.com/seunghwak)

URL: https://github.com/rapidsai/cugraph/pull/3639