[DOC]: Fix invalid links and add materials to notebooks (#4002)
I'm new to CuGraph repository and would like to contribute further in the future. So I start with minor tasks with docs to familiarize myself with the contribution process in this repository.

Here is what I have done in this PR:
- Fixed some invalid links in the documentation.
- Corrected and added formulas related to centrality.
- Added more references to the centrality to ensure consistency.

Also, the images located at https://github.com/rapidsai/cugraph/tree/main/python/cugraph cannot be displayed because it is symbolic link to the repository's README file. But it's not a major issue.

Authors:
  - Huiyu Xie (https://github.com/huiyuxie)
  - Ralph Liu (https://github.com/nv-rliu)
  - Naim (https://github.com/naimnv)
  - Brad Rees (https://github.com/BradReesWork)

Approvers:
  - Don Acosta (https://github.com/acostadon)
  - Brad Rees (https://github.com/BradReesWork)

URL: https://github.com/rapidsai/cugraph/pull/4002