Fix bad Leiden result (#4376)
Once the termination criteria is met, the last level of the dendrogram contains a partitioning that reduces the overall modularity.  This update overwrites the final dendrogram level with a k to k mapping. 

This bug had a consequence of the numbering of the flattened Leiden partitions being non-consecutive integers.  This should now be resolved.

Closes #4368

Authors:
  - Chuck Hastings (https://github.com/ChuckHastings)

Approvers:
  - Seunghwa Kang (https://github.com/seunghwak)

URL: https://github.com/rapidsai/cugraph/pull/4376